+++ library/all-exceptions/index.ts
@@ -1,0 +1,2 @@
+export * from './all-exception.filter';
+export * from './interface/exception-message.interface';
