--- library/index.ts
@@ -1,2 +1,0 @@
-export * from './all-exceptions/all-exception.filter';
-export * from './all-exceptions/interface/exception-message.interface';
